Rename WS to Gamewebsocket
