+++ server/src/main.rs
@@ -1,25 +1,16 @@
 mod api;
 mod error;
+mod gamewebsocket;
 mod server;
 mod session;
 mod ws;
 
-
-use actix::{Actor};
-
-
+use actix::Actor;
 
 use actix_web::{web, App, HttpServer};
-use anyhow::{Result};
+use anyhow::Result;
 
 use cult_common::*;
-
-
-
-
-
-
-
 
 #[actix_web::main]
 async fn main() -> Result<()> {
@@ -32,7 +23,7 @@ async fn main() -> Result<()> {
     let server = HttpServer::new(move || {
         App::new()
             .app_data(web::Data::new(server.clone()))
-            .route("/ws", web::get().to(ws::start_ws))
+            .route("/ws", web::get().to(gamewebsocket::start_ws))
             .service(api::game_info)
     })
     .bind(addr)?
